+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [ConVista Faktor Zehn GmbH](http://www.faktorzehn.at/).
 ## [unreleased]
 ### Fixed
 - jenkins.job: added job state ABORTED
+### Changed
+- projects: new property sonarproject to configure one global project for metrics and qualitygate for this project
 
 ## [1.1.8] - 2017-05-18
 ### Fixed